Taxes_Managements/Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
cryptography
